compost tracking
Implements compost tracking for farming patches using a similar concept as the Examine plugin.

The plugin watches for composting and inspect actions, then stores a pending compost which is finalized by chat messages. In order to distinguish between multiple compost actions queued, the tracker also checks for adjacency to the patch and has a timeout of 30 seconds.

At the time, the plugin does not display the information to the user, since I know there is ongoing discussion about how to do so. Hopefully this will help to drive that forward now that the detection portion is completed.