IconTextField: select hovered suggestion
Substance had a hover effect that it would show here, but FlatLaf (and
Basic) do not, and quite a bit of code is needed for this sort of thing.
FlatLaf's (and Basic's) ComboBox instead selects the hovered list item
and has a separate mouse listener to actually update the ComboBox's
model. This just does the first part, making it so they list filters
upon hovering a suggestion, but I don't think that is bad for our use
cases.